moved branding.constant.ts -> branding.ts
